ENH: Add named analog channels and hardware DIO
* All channels (digital and analog) now support user-provided names that
will be used in the callbacks. This facilitates third-party data
acquisition code that may be helped by knowing which channels the
callback pertains to.
* Added hardware timed digital inputs. Includes an edge detection
coroutine to extract epoch times. The edge detection supports debouncing
of the signal as well.
* The samples generated callback helper now provides the current offset
in the buffer in addition to the number of samples available for writing
in the buffer.